rename: accordion을 TermsOfUse 안으로 이동
