+++ src/pages/Register/TermsOfUse/Accordion/index.tsx
