+++ src/pages/Register/TermsOfUse/Accordion/styles.ts
